Hi Team,
I modified the merged file to remove all blank lines and changed the values of the arrays in the main method to be dynamically generated using random numbers. I previously implemented it but the latest merged file made the array elements static. Its fixed now.

I also created a new directory in the project to store the screenshot of the running of each methods

***Important: Please write comments next to each line in the method you created***

Also Please remove your self- reflection comments from the code and write something like i have done next to my name. You will submit your self reflection as a comment when submitting the link on Canvas.

--Thanks,
--Richmond.